Remove title from index.
Due to https://github.com/quarto-dev/quarto-cli/issues/3673 it is not possible at the moment to conditionally have a title for the website profile (as we would like) and not have a title at all for the book profile (as we would like). When this issue is resolved, we might be able to get back to this.